requirements.txt
torchaudio==2.0.1
torch==2.0.0
pytorch-lightning==1.9.4
tqdm==4.65.0
scikit-learn==1.2.2
pandas==1.5.3
pyyaml==6.0
timm==0.6.12
torchtoolbox==0.1.8.2
scipy
librosa
kaggle
icecream
matplotlib
seaborn
plotly